refactor: recordlist에서 my-records로 변경
